# api/openapi.yaml
openapi: 3.1.0
info:
  title: ImageSilo API
  version: 0.1.0
  description: ImageSilo V1 的唯一 HTTP 契约来源。
servers:
  - url: /
paths:
...
  /api/v1/images:
    get:
      operationId: listImages
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: cursor
          in: query
          schema: { type: string }
        - name: q
          in: query
          description: 按原始文件名、标准路径 UUID、历史路径或 SHA-256 搜索
          schema: { type: string, maxLength: 200 }
        - name: visibility
          in: query
          schema: { $ref: '#/components/schemas/Visibility' }
        - name: format
          in: query
          schema: { type: string, enum: [jpeg, png, webp, gif] }
        - name: uploadedVia
          in: query
          schema: { type: string, enum: [admin, api_token, import] }
        - name: createdFrom
          in: query
          schema: { type: string, format: date-time }
        - name: createdTo
          in: query
          schema: { type: string, format: date-time }
        - name: minBytes
          in: query
          schema: { type: integer, format: int64, minimum: 0 }
        - name: maxBytes
          in: query
          schema: { type: integer, format: int64, minimum: 0 }
        - name: minWidth
          in: query
          schema: { type: integer, minimum: 0 }
        - name: maxWidth
          in: query
          schema: { type: integer, minimum: 0 }
        - name: minHeight
          in: query
          schema: { type: integer, minimum: 0 }
        - name: maxHeight
          in: query
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: 使用稳定 keyset cursor 的分页图片列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      operationId: uploadImage
      security:
        - sessionCookie: []
        - bearerToken: ['images:upload']
      parameters:
        - $ref: '#/components/parameters/CSRFTokenForSession'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                visibility:
                  $ref: '#/components/schemas/Visibility'
      responses:
        '201':
          description: 图片上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/images/{imageId}/visibility:
    patch:
      operationId: updateImageVisibility
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ImageId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateImageVisibilityRequest'
      responses:
        '204':
          description: 可见性已提交到 SQLite 并同步更新交付索引
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/images/{imageId}/thumbnail:
    get:
      operationId: getImageThumbnail
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ImageId'
      responses:
        '200':
          description: 可重建的后台 JPEG 预览缩略图
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: imagesilo_session
    bearerToken:
      type: http
      scheme: bearer
  parameters:
    ImageId:
      name: imageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: 必须与当前 Session 绑定的 imagesilo_csrf Cookie 完全一致。
      schema:
        type: string
        pattern: '^isc_[A-Za-z0-9_-]{43}$'
    CSRFTokenForSession:
      name: X-CSRF-Token
      in: header
      required: false
      description: 使用管理员 Session 上传时必填；使用 Bearer Token 时忽略。
      schema:
        type: string
        pattern: '^isc_[A-Za-z0-9_-]{43}$'
  schemas:
    Visibility:
      type: string
      enum: [public, private]
...
    Image:
      type: object
      required: [id, originalName, mimeType, extension, width, height, sourceSize, storedSize, sourceSha256, storedSha256, processingSummary, visibility, uploadedVia, standardUrl, thumbnailUrl, createdAt]
      properties:
        id: { type: string, format: uuid }
        originalName: { type: string }
        mimeType: { type: string, enum: [image/jpeg, image/png, image/webp, image/gif] }
        extension: { type: string, enum: [.jpg, .png, .webp, .gif] }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
        sourceSize: { type: integer, format: int64, minimum: 1 }
        storedSize: { type: integer, format: int64, minimum: 1 }
        sourceSha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        storedSha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        processingSummary: { $ref: '#/components/schemas/ProcessingSummary' }
        visibility: { $ref: '#/components/schemas/Visibility' }
        uploadedVia: { type: string, enum: [admin, api_token, import] }
        standardUrl: { type: string }
        thumbnailUrl: { type: string }
        createdAt: { type: string, format: date-time }
    ImageList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Image' }
        nextCursor: { type: string }
...
    UpdateImageVisibilityRequest:
      type: object
      additionalProperties: false
      required: [visibility]
      properties:
        visibility: { $ref: '#/components/schemas/Visibility' }
...
    ProcessingSummary:
      type: object
      required: [action, sourceFormat, storedFormat, preserved, compressionEnabled, conversionEnabled]
      properties:
        action: { type: string, enum: [preserve, compress, convert] }
        sourceFormat: { type: string, enum: [jpeg, png, webp, gif] }
        storedFormat: { type: string, enum: [jpeg, png, webp, gif] }
        preserved: { type: boolean }
        compressionEnabled: { type: boolean }
        conversionEnabled: { type: boolean }
        compressionRejected: { type: boolean }
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
  responses:
    BadRequest:
      description: 请求无效
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: 缺少或认证信息无效
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: 已认证但 Scope 不足
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: 资源不存在
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ServiceUnavailable:
      description: 服务未就绪或图片解码安全门满载
      headers:
        Retry-After: { schema: { type: integer, minimum: 1 } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }